Setting up Docker correctly. Tested on Mac.
For whatever reason tho, it doesn't show up in the terminal like it usually does.
You have to go to Localhost:8000 manually.
Would recomend getting docker extension on vscode.
Going over to the containers running and rightclicking it will allow you to attach a shell and also on up in browser
pretty dope.